Updated layout and structure
This commit updates the site layout, adding a navbar at the top
and update bootstrap to the latest version (using bootstrap CDN).

It also introduces a mechanism to automatically update the download
links using a jekyll "post" with the right front-matter. See existing
post _posts/2013-03-18-v2.5.0.md for an example.